% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{criteriaTable}
\alias{criteriaTable}
\title{Criteria Table}
\usage{
criteriaTable(model)
}
\arguments{
\item{model}{The model you want to evaluate.}
}
\description{
This function accepts a linear model as an input
and returns a bunch of useful values for model selection.
Included values: df, r.squared, SSE, RMSE, RMSP, AIC and BIC.
}
\examples{
criteriaTable()
}
\keyword{aic}
\keyword{bic}
\keyword{df}
\keyword{model}
\keyword{rmse}
\keyword{rmsp}
\keyword{sse}

