% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{latexLm}
\alias{latexLm}
\title{Latex Linear Model}
\usage{
latexLm(model)
}
\arguments{
\item{model}{The model to print as Latex.}
}
\description{
This function accepts a linear model as an input
and cats a Latex string. You can then paste this into
a Latex or RMarkdown file.
}
\examples{
latexLm()
}
\keyword{latex}
\keyword{model}

