% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{pairwiseCor}
\alias{pairwiseCor}
\title{Pairwise Correlation}
\usage{
pairwiseCor(dataFrame)
}
\arguments{
\item{dataFrame}{The dataframe you want correlations from.}
}
\description{
This function accepts a dataframe as input and returns
a dataframe with the pairwise absolute correlation and
regular correlation. Make sure to give it numeric input.
}
\examples{
pairwiseCor()
}
\keyword{correlation}
\keyword{pairwise}
\keyword{tidy}

