% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rmsp}
\alias{rmsp}
\title{RMSP}
\usage{
rmsp(model)
}
\arguments{
\item{model}{A model object.}
}
\description{
This function accepts a model object as input and returns
the 'RMSP' (root-mean-square-PRESS) statistic for that model.
}
\examples{
rmsp()
}
\keyword{model}
\keyword{press}
\keyword{rmsp}

